requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
prisma==0.15.0
python-dotenv==1.0.0
langchain-openai>=0.0.5
langchain-pinecone>=0.0.3
langchain-core>=0.1.18
langchain-community>=0.0.16
langsmith>=0.1.4
aiohttp>=3.9.0 